clean up logout code
