+++ src/features/auth/useSavedAuth.ts
@@ -1,10 +1,9 @@
 import assert from 'assert';
 
-import { client } from 'lib/gql/client';
 import { atom, selector, useRecoilValue } from 'recoil';
 
 import type { ProviderType } from './store';
-import { getAuthToken, setAuthToken } from './token';
+import { setAuthToken } from './token';
 
 export enum EConnectorNames {
   Injected = 'injected',
@@ -18,17 +17,8 @@ export interface IAuth {
   authTokens: { [k: string]: string | undefined };
 }
 
-const logout = async (): Promise<boolean> => {
-  const { logoutUser } = await client.mutate(
-    { logoutUser: { id: true } },
-    { operationName: 'logout' }
-  );
-  return !!logoutUser?.id;
-};
-
 const updateToken = ({ address, authTokens }: IAuth) => {
   const token = address && authTokens[address];
-  if (!token && getAuthToken(false)) logout();
   setAuthToken(token);
 };
 
@@ -36,6 +26,7 @@ const AUTH_STORAGE_KEY = 'capeAuth';
 
 const saveAuth = (auth: IAuth) =>
   localStorage.setItem(AUTH_STORAGE_KEY, JSON.stringify(auth));
+
 const getSavedAuth = (): IAuth => {
   try {
     const auth = localStorage.getItem(AUTH_STORAGE_KEY);
